+++ lib/lists.dart
@@ -7,6 +7,7 @@ part "src/bit_list.dart";
 part "src/filled_list.dart";
 part "src/grouped_range_list.dart";
 part "src/range_list.dart";
+part "src/sparse_bool_list.dart";
 part "src/sparse_list.dart";
 part "src/step_list.dart";
 part "src/wrapped_list.dart";